Remove `mindplay/composer-locator` package.
Removing the package again, as it unfortunately only works _after_ the autoloader has been dumped, which makes it useless in our case (where we want to use it to control autoloader generation).

Falling back to just assuming that `wp-cli/wp-cli` is installed in the `vendor/wp-cli/wp-cli` folder, as the root package can still override this value.